Big refactor: Because word_matcher is a tool, it should not contain a domain module, and only focus on the infrastructure module.
