+++ README.md
@@ -1,2 +1,2 @@
 
-![Alt text](https://raw.githubusercontent.com/miguelemosreverte/word_matcher/0.0.0/Leafin%20WordMatcher%20Readme.svg?token=ACF2PCBPKUYXR7C3IPYKIMTAY7HXU)
+![Alt text](https://raw.githubusercontent.com/miguelemosreverte/word_matcher/main/Leafin%20WordMatcher%20Readme.svg?token=ACF2PCG7NHR62MGKED7M2WLAY7IQU)
