+++ build.sbt
@@ -1,6 +1,6 @@
 lazy val `word_matcher` =
   (project in file("."))
-    .aggregate(`domain`, `application`)
+    .aggregate(`domain`, `application`, `infrastructure`)
     .settings(
       assembly / mainClass := Some("leapfin.lemos.word_matcher.Main"),
       assembly / assemblyJarName := "word_matcher.jar"
@@ -11,9 +11,9 @@ lazy val `domain` = project in file(".") / "modules" / "domain"
 lazy val `application` =
   (project in file(
     "."
-  ) / "modules" / "application" dependsOn (`domain`, `infrastructure`))
+  ) / "modules" / "application" dependsOn (`domain` % "compile->compile;test->test",
+  `infrastructure` % "compile->compile;test->test"))
 
 lazy val `infrastructure` = project in file(".") / "modules" / "infrastructure"
 
 ThisBuild / Test / testOptions += Tests.Argument("-oD")
-ThisBuild / IntegrationTest / testOptions += Tests.Argument("-oD")
