+++ modules/infrastructure/src/main/scala/leapfin/lemos/word_matcher/interpreter/Config.scala
