+++ modules/infrastructure/src/main/scala/leapfin/lemos/word_matcher/interpreter/WordMatcher.scala
@@ -2,29 +2,14 @@ package leapfin.lemos.word_matcher.interpreter
 
 import akka.actor.ActorSystem
 import leapfin.infrastructure.stream.utils.search.SlidingWindowSearch
-import leapfin.lemos.word_matcher.interpreter.WordMatcher.{
-  Logger,
-  MatchResult,
-  successLogger
-}
-import leapfin.lemos.word_matcher.algebra.{
-  Status,
-  WordMatcher => WordMatcherContract
-}
+import leapfin.infrastructure.stream.utils.search.SlidingWindowSearch._
+import leapfin.lemos.word_matcher.Status
+import leapfin.lemos.word_matcher.algebra.WordMatcher.MatchResult
+import leapfin.lemos.word_matcher.algebra.{WordMatcher => WordMatcherContract}
 
 import scala.concurrent.ExecutionContext
 import scala.concurrent.duration.FiniteDuration
 import scala.language.postfixOps
-
-object WordMatcher {
-  type MatchResult = Either[Status.Failure, Status.Success]
-  type Logger = MatchResult => Unit
-  val emptyLogger: Logger = _ => ()
-  val successLogger: Logger = {
-    case Left(timeout)  => ()
-    case Right(success) => println(s"Found the word! -- $success")
-  }
-}
 
 class WordMatcher(
     config: Config,
@@ -47,7 +32,7 @@ class WordMatcher(
       timeout
     ) match {
       case Left(value) =>
-        Left(Status.Timeout(value.byteCount))
+        Left(Status.NotFound(value.byteCount))
       case Right(value) =>
         Right(Status.Success(value.elapsedTime, value.byteCount))
     }
