+++ modules/infrastructure/src/test/scala/leapfin/lemos/word_matcher/naive_spec/WordMatcherSpec.scala
@@ -1,11 +1,10 @@
 package leapfin.lemos.word_matcher.naive_spec
 
-import leapfin.lemos.word_matcher.Mock
 import leapfin.lemos.word_matcher.algebra.{
-  WordMatcherSpec => WordMatcherAbstractSpec
+  WordMatcherSpec => WordMatcherContract
 }
 
 class WordMatcherSpec
-    extends WordMatcherAbstractSpec(
+    extends WordMatcherContract(
       Mock
     )
